Rename spark facets according to new naming
