+++ integration/spark/shared/src/main/java/io/openlineage/spark/agent/facets/GcpLineageJobFacet.java
@@ -13,17 +13,23 @@ import java.net.URI;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public class GCPDataprocRunFacet implements OpenLineage.RunFacet {
+public class GcpLineageJobFacet implements OpenLineage.JobFacet {
 
   private final URI _producer;
   private final URI _schemaURL;
+  private final Boolean _deleted;
   @JsonAnySetter private final Map<String, Object> additionalProperties;
 
-  public GCPDataprocRunFacet(Map<String, Object> properties) {
+  public GcpLineageJobFacet(Map<String, Object> properties) {
+    this(properties, null);
+  }
+
+  public GcpLineageJobFacet(Map<String, Object> properties, Boolean _deleted) {
     this._producer = Versions.OPEN_LINEAGE_PRODUCER_URI;
     this._schemaURL =
         URI.create(
-            "https://openlineage.io/spec/facets/1-0-0/GcpDataprocRunFacet.json#/$defs/GcpDataprocRunFacet");
+            "https://openlineage.io/spec/facets/1-0-0/GcpLineageJobFacet.json#/$defs/GcpLineageJobFacet");
+    this._deleted = _deleted;
     this.additionalProperties = new LinkedHashMap<>(properties);
   }
 
@@ -35,6 +41,11 @@ public class GCPDataprocRunFacet implements OpenLineage.RunFacet {
   @Override
   public URI get_schemaURL() {
     return this._schemaURL;
+  }
+
+  @Override
+  public Boolean get_deleted() {
+    return this._deleted;
   }
 
   @JsonAnyGetter
