+++ spec/registry/gcp/dataplex/facets/GcpLineageJobFacet.json
@@ -1,8 +1,8 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://openlineage.io/spec/facets/1-0-0/DataplexJobFacet.json",
+  "$id": "https://openlineage.io/spec/facets/1-0-0/GcpLineageJobFacet.json",
   "$defs": {
-    "DataplexJobFacet": {
+    "GcpLineageJobFacet": {
       "allOf": [
         {
           "$ref": "https://openlineage.io/spec/2-0-2/OpenLineage.json#/$defs/JobFacet"
@@ -36,8 +36,8 @@
   },
   "type": "object",
   "properties": {
-    "dataplex": {
-      "$ref": "#/$defs/DataplexJobFacet"
+    "gcp_lineage": {
+      "$ref": "#/$defs/GcpLineageJobFacet"
     }
   }
 }
